api-gateway/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pika==1.3.2
pydantic==2.11.9
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
